Fix one possible cause of disabling lockstep mode completely freezing the game.
If lockstep was enabled when exactly one graphical frame was queued (any time in a menu, or a 50-50 chance during gameplay with default FPS caps), outstanding_gfps would end up negative. lockstep would partially disable itself, returning control to libgraphics, but not get far enough to replace its fake SDL_GetTicks function with the real one. This would cause libgraphics to wait forever for the next frame, as time could not pass while lockstep was in this state, and lockstep would not get a chance to disable its clock hooks until the clock advanced, which would never happen.

This commit also resets the FPS calculation data when lockstep is disabled.